requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.4
Faker==15.3.3
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
structlog==22.3.0
tomli==2.0.1
urllib3==1.26.13
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.4
Faker==15.3.3
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
structlog==22.3.0
tomli==2.0.1
urllib3==1.26.13
